+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/assets/domain/importpoc/service/PolicyServiceImplTest.java
@@ -19,9 +19,17 @@
 package org.eclipse.tractusx.traceability.assets.domain.importpoc.service;
 
 
-import org.eclipse.tractusx.irs.edc.client.policy.AcceptedPoliciesProvider;
-import org.eclipse.tractusx.irs.edc.client.policy.AcceptedPolicy;
+import assets.importpoc.PolicyResponse;
+import org.eclipse.tractusx.irs.edc.client.policy.Constraint;
+import org.eclipse.tractusx.irs.edc.client.policy.Constraints;
+import org.eclipse.tractusx.irs.edc.client.policy.Operator;
+import org.eclipse.tractusx.irs.edc.client.policy.OperatorType;
+import org.eclipse.tractusx.irs.edc.client.policy.Permission;
 import org.eclipse.tractusx.irs.edc.client.policy.Policy;
+import org.eclipse.tractusx.irs.edc.client.policy.PolicyType;
+import org.eclipse.tractusx.traceability.assets.domain.base.IrsRepository;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.IrsPolicyResponse;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Payload;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -35,18 +43,18 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
+
 @ExtendWith(MockitoExtension.class)
 class PolicyServiceImplTest {
     @InjectMocks
     private PolicyServiceImpl policyService;
     @Mock
-    private AcceptedPoliciesProvider acceptedPoliciesProvider;
+    private IrsRepository irsRepository;
 
     @BeforeEach
     public void testSetup() {
-        policyService = new PolicyServiceImpl(acceptedPoliciesProvider);
+        policyService = new PolicyServiceImpl(irsRepository);
     }
-
 
 
     @Test
@@ -56,17 +64,35 @@ class PolicyServiceImplTest {
         // GIVEN
         String policyId = "policy123";
         OffsetDateTime createdOn = OffsetDateTime.parse("2023-07-03T16:01:05.309Z");
-        List<AcceptedPolicy> acceptedPolicies = List.of(
-                new AcceptedPolicy(new Policy("policy123", createdOn,null, null),null));
+        List<IrsPolicyResponse> acceptedPolicies = List.of(
+                IrsPolicyResponse.builder()
+                        .validUntil(OffsetDateTime.now())
+                        .payload(
+                                Payload.builder()
+                                        .policyId(policyId)
+                                        .policy(
+                                                Policy.builder()
+                                                        .createdOn(createdOn)
+                                                        .permissions(List.of(
+                                                                Permission.builder()
+                                                                        .action(PolicyType.USE)
+                                                                        .constraint(Constraints.builder()
+                                                                                .and(List.of(new Constraint("", new Operator(OperatorType.EQ), "")))
+                                                                                .or(List.of(new Constraint("", new Operator(OperatorType.EQ), "")))
+                                                                                .build())
+                                                                        .build()))
+                                                        .build())
+                                        .build())
+                        .build());
 
         // WHEN
-        when(acceptedPoliciesProvider.getAcceptedPolicies()).thenReturn(acceptedPolicies);
-        List<Policy> result = policyService.getAllPolicies();
+        when(irsRepository.getPolicies()).thenReturn(acceptedPolicies);
+        List<PolicyResponse> allPolicies = policyService.getAllPolicies();
 
         // THEN
-        assertNotNull(result);
-        assertEquals(policyId, result.get(0).getPolicyId());
-        assertEquals(createdOn, result.get(0).getCreatedOn());
+        assertNotNull(allPolicies);
+        assertEquals(policyId, allPolicies.get(0).policyId());
+        assertEquals(createdOn, allPolicies.get(0).createdOn());
     }
 
 }