+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ### Added
 - #515 Service Unavailable Response on Notification failure
+- #536 Added import state PUBLISHED_TO_CX in frontend
 - #420 add /contracts api to fetch contract agreement information from EDC for assets
 - Added a step to the pull-request-backend.yml which checks if the pom.xml(root) properties have some versions ending with -SNAPSHOT
 
@@ -20,6 +21,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #625 increased height of tables
 - #423 Moved errorMessages from investigation/alert to notification list
 - Updated COMPATIBILITY.md matrix adding release C-X 24.3 and 23.12
+- #536 rework /policies to respond with policies from the IRS policy store
 
 ### Removed
 - #625 Removed the header and breadcrumbs section from app layout